requirements.txt
# Asistente Telefónico con IA - Dependencias para Producción

# OpenAI API
openai>=1.12.0

# Variables de entorno
python-dotenv>=1.0.0

# Web Framework
flask>=3.0.0
gunicorn>=21.2.0

# Twilio para WhatsApp
twilio>=9.0.0

# Reconocimiento de voz
SpeechRecognition>=3.10.0

# Síntesis de voz
gTTS>=2.5.0

# HTTP requests
requests>=2.31.0

# Audio processing (opcional - solo si usas audio local)
# pyttsx3>=2.90
# PyAudio>=0.2.14
